requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
matplotlib>=3.9
pyyaml>=6.0
joblib>=1.4